music_generator/requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# For REAL singing with music (Eleven Labs Music API)
elevenlabs>=1.0.0

# Optional: For AI-assisted lyrics generation
# Uncomment if you want to use the --ai feature in interactive_singing.py
# openai>=1.0.0
